Revert "Test on logs"
This reverts commit a133ba0cd33dfdb07b54e0f8491284cd2ef4a358.
